r: Commander => EncounterCommander
